@@ -1,51 +1,29 @@
 {
   "name": "include-type",
-  "version": "0.1.1",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@fortawesome/react-fontawesome": "^0.1.14",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.60",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.1.9",
+    "@testing-library/jest-dom": "^5.15.0",
+    "@testing-library/react": "^11.2.7",
+    "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^26.0.24",
-    "@types/node": "^16.0.0",
-    "@types/react": "^17.0.13",
-    "@types/react-dom": "^17.0.8",
-    "@types/react-router": "^5.1.15",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-scroll": "^1.8.2",
-    "bootstrap": "^5.0.2",
+    "@types/node": "^12.20.37",
+    "@types/react": "^17.0.34",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-scroll": "^1.8.3",
     "react": "^17.0.2",
-    "react-bootstrap": "^1.6.1",
-    "react-calendar": "^3.4.0",
-    "react-country-region-selector": "^3.1.0",
+    "react-country-region-selector": "^3.4.0",
     "react-dom": "^17.0.2",
-    "react-easy-crop": "^3.5.2",
+    "react-easy-crop": "^3.5.3",
     "react-phone-input-2": "^2.14.0",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-scroll": "^1.8.2",
-    "smooth-scrollbar-react": "^2.2.0",
-    "typescript": "^4.3.5",
-    "web-vitals": "^2.1.0",
-    "workbox-background-sync": "^6.1.5",
-    "workbox-broadcast-update": "^6.1.5",
-    "workbox-cacheable-response": "^6.1.5",
-    "workbox-core": "^6.1.5",
-    "workbox-expiration": "^6.1.5",
-    "workbox-google-analytics": "^6.1.5",
-    "workbox-navigation-preload": "^6.1.5",
-    "workbox-precaching": "^6.1.5",
-    "workbox-range-requests": "^6.1.5",
-    "workbox-routing": "^6.1.5",
-    "workbox-strategies": "^6.1.5",
-    "workbox-streams": "^6.1.5"
+    "react-router-dom": "^6.0.2",
+    "react-scripts": "^4.0.3",
+    "react-scroll": "^1.8.4",
+    "typescript": "^4.4.4",
+    "web-vitals": "^1.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
