@@ -4,11 +4,9 @@ import "../css/first-landing-section-css/BrandName.css";
 export default function BrandName() {
   return (
     <div className="brand_container py-4 px-4">
-      <h1>
-        <span className="name_1">#include &lt;</span>
-        <span className="name_2">TYPE</span>
-        <span className="name_1">&gt;</span>
-      </h1>
+      <span className="name_1">#include &lt;</span>
+      <span className="name_2">TYPE</span>
+      <span className="name_1">&gt;</span>
     </div>
   );
 }
